feat(protocol): add canonical Category type + catalog to arclib (Stage 4 step 1)
First step of moving the MyClerk opcode catalog into arclib so every app
(core, mayaservices, ArcShell, ArcMail, the terminal) uses one source of
truth and the per-project mirroring stops. Adds protocol/category.go with
the Category type, all category constants (0x00-0x31 + 0xFF) and names.
OpCode.Category() now returns Category (was uint8) and NewOpCode takes a
Category; the lone internal caller (arcmail) and tests are updated.

+++ protocol/gen_vectors_test.go
@@ -245,7 +245,7 @@ func TestGenerateProtocolVectors(t *testing.T) {
 		vectors.OpCodes = append(vectors.OpCodes, opcodeVector{
 			Code:     uint16(op),
 			Name:     op.String(),
-			Category: op.Category(),
+			Category: uint8(op.Category()),
 			Op:       op.Operation(),
 		})
 	}
